Add widget.c
