+++ include/LCUI/gui/widget/button.h
@@ -45,6 +45,8 @@ LCUI_API void Button_SetTextW( LCUI_Widget w, const wchar_t *wstr );
 
 LCUI_API void Button_SetText( LCUI_Widget w, const char *str );
 
+LCUI_API void LCUIWidget_AddButton( void );
+
 LCUI_END_HEADER
 
 #endif
