+++ include/LCUI/gui/widget/sidebar.h
@@ -42,10 +42,12 @@
 
 LCUI_BEGIN_HEADER
 
-LCUI_API LCUI_Widget SideBar_AppendItem( LCUI_Widget sidebar, const wchar_t *id, 
-				const wchar_t *icon, const wchar_t *text );
+LCUI_API LCUI_Widget SideBar_AppendItem( LCUI_Widget sidebar,
+					 const wchar_t *id,
+					 const wchar_t *icon,
+					 const wchar_t *text );
 
-void LCUIWidget_AddSideBar(void);
+LCUI_API void LCUIWidget_AddSideBar( void );
 
 LCUI_END_HEADER
