+++ include/LCUI/gui/widget/textview.h
@@ -40,8 +40,6 @@
 #ifndef LCUI_TEXTVIEW_H
 #define LCUI_TEXTVIEW_H
 
-#include <LCUI/font.h>
-
 LCUI_BEGIN_HEADER
 
 /** 设定与标签关联的文本内容 */
@@ -54,6 +52,8 @@ LCUI_API void TextView_SetLineHeight( LCUI_Widget w, int height );
 LCUI_API void TextView_SetTextAlign( LCUI_Widget w, int align );
 
 LCUI_API void TextView_SetAutoWrap( LCUI_Widget w, LCUI_BOOL autowrap );
+
+LCUI_API void LCUIWidget_AddTextView( void );
 
 LCUI_END_HEADER
