+++ src/gui/widget.c
@@ -1,7 +1,7 @@
 ﻿/* ***************************************************************************
- * widget.h -- GUI widget APIs.
+ * widget.c -- GUI widget APIs.
  *
- * Copyright (C) 2015-2017 by Liu Chao <lc-soft@live.cn>
+ * Copyright (C) 2017 by Liu Chao <lc-soft@live.cn>
  *
  * This file is part of the LCUI project, and may only be used, modified, and
  * distributed under the terms of the GPLv2.
@@ -20,9 +20,9 @@
  * ****************************************************************************/
 
 /* ****************************************************************************
- * widget.h -- GUI部件操作相关的函数接口。
+ * widget.c -- GUI 部件相关功能的初始化与释放接口
  *
- * 版权所有 (C) 2015-2017 归属于 刘超 <lc-soft@live.cn>
+ * 版权所有 (C) 2017 归属于 刘超 <lc-soft@live.cn>
  *
  * 这个文件是LCUI项目的一部分，并且只可以根据GPLv2许可协议来使用、更改和发布。
  *
@@ -37,20 +37,41 @@
  * 没有，请查看：<http://www.gnu.org/licenses/>.
  * ****************************************************************************/
 
-#ifndef LCUI_WIDGET_BUILD_H
-#define LCUI_WIDGET_BUILD_H
+#include <LCUI_Build.h>
+#include <LCUI/LCUI.h>
+#include <LCUI/gui/widget.h>
+#include <LCUI/gui/widget/textview.h>
+#include <LCUI/gui/widget/textcaret.h>
+#include <LCUI/gui/widget/textedit.h>
+#include <LCUI/gui/widget/button.h>
+#include <LCUI/gui/widget/sidebar.h>
+#include <LCUI/gui/widget/scrollbar.h>
 
-#include <LCUI/graph.h>
-#include <LCUI/thread.h>
-#include <LCUI/gui/widget_base.h>
-#include <LCUI/gui/widget_task.h>
-#include <LCUI/gui/widget_paint.h>
-#include <LCUI/gui/widget_prototype.h>
-#include <LCUI/gui/widget_event.h>
-#include <LCUI/gui/widget_style.h>
+void LCUI_InitWidget( void )
+{
+	LCUIWidget_InitTasks();
+	LCUIWidget_InitEvent();
+	LCUIWidget_InitPrototype();
+	LCUIWidget_InitStyle();
+	LCUIWidget_InitRenderer();
+	LCUIWidget_InitImageLoader();
+	LCUIWidget_AddTextView();
+	LCUIWidget_AddButton();
+	LCUIWidget_AddSideBar();
+	LCUIWidget_AddTScrollBar();
+	LCUIWidget_AddTextCaret();
+	LCUIWidget_AddTextEdit();
+	LCUIWidget_InitBase();
+}
 
-void LCUI_InitWidget( void );
-
-void LCUI_ExitWidget( void );
-
-#endif
+void LCUI_ExitWidget( void )
+{
+	LCUIWidget_ExitTasks();
+	LCUIWidget_DestroyRoot();
+	LCUIWidget_ExitEvent();
+	LCUIWidget_ExitStyle();
+	LCUIWidget_ExitPrototype();
+	LCUIWidget_ExitRenderer();
+	LCUIWidget_ExitImageLoader();
+	LCUIWidget_ExitBase();
+}
